Added chime switch support for DSC
DSC panels do not appear to provide a way to query whether the chime is enabled or disabled so the chime switch will now use HA to restore the entity's state (RestoreEntity) to try and ensure the state matches on HA startup.  It will still be possible for this state to be out of sync on startup if the chime has been toggled while HA is not running.