Allow separate address for discovery
The discovery port configuration option now accepts an optional hostname/IP.  This allows people using a proxy like AlarmServer or DscServer to configure the discovery endpoint as the real EVL device but router their TPI connection through the proxy.  If a hostname is required, the `Discovery Port` setting must have the format `hostname:port`.  For example: `127.0.0.1:8080` or `localhost:80`.