Use default code set on alarm_panel_entity.
HA `alarm_control_panel` entities can be configured with a default alarm code. If set, `envisalink_new` will now use this code instead of what is stored in the config entry.  If it is not set on the `alarm_control_panel` entity, the code configured in the config entry will be used.  Warnings will be logged if a code is set in both places or if a code is set only in the config entry but not.